requirements.txt
psutil
GPUtil
pystray
Pillow
PySide6
